rework: Reworked Actions Resize/Reshape
Actions in ui module is now also spilt into submodules and moved to
use plugin system.

Actions in core is now implemented as a que, this helps to buffer
actions one after the other.

Removed OnSpawnBase event, added OnReshapeBase event. It was
never inteded to have more than 1 base, and we want to have a base
selection on application start, so on startup we que a Reshape and
a Resize action.