feat: Updated Base Actions
Renamed OnReshapeBase Event, now it is known as OnSpawnBase, this
action should be called only once. Since Base is a now a Resource,
we insert it on this event.

Base also holds a Handle to its Mesh instead of the Mesh data.

Removed OnResizeBase Event since in the future it will be replaced by
a new action, which lets the user to reposition the nodes of the Base.