feat: Updated picking to restore original matcap
Previously object picking always overwritten the original matcap of
the selected object. From now on the original one is stored inside
SelectedMesh component and will be restored when the its deselected.

Also changed picking to run on Pressed signal instead of Released,
so that it's matches other parts of the application.
(e.g. button press)