Reworked mesh picking plugin
Now only meshes with Pickable component can be picked. Pickable is
also a list of mesh types, that are meant to picked.
(For now its only holds Node, which is will be used in EditBase action)

Removed old logic of tagging picked meshes with SelectedMesh component.
This was used to change the color/material of the picked mesh, which
made custom coloring/material hard. Instead now each system have to
read the pointer event and handle the color/material change.

Updated documentations to reflect the changes.