Added initial implementation of EditBase action
In the EditBase action the mesh of the Base gets hidden and an
editor is displayed. Interactable nodes are spawned, each of them
can be relocated along X or Z axis, moving the other node along the
same axis.

Additionally info displayed at the bottom about the current state.
Users can leave the editor by clicking on the confirm button.